+++ HISTORY.md
@@ -5,7 +5,8 @@
 ### UI Changes
 
 * Remove confusing `activate` and `deactivate` commands in favour of the `use` command. The previous “reactivation” behaviour is now mapped to `link --all`.
-* `snafu upgrade self` to perform in-place upgrade without manually downloading the installer.
+* Add `snafu upgrade self` to perform in-place upgrade without manually downloading the installer.
+* Add `snafu download <version>` to download installer of given Python version without installing.
 * More complete help messages are provided to command arguments.
 
 ### Behavioural Changes